feat: implement simple cookie parsing

+++ rest-assert-core/src/main/java/com/github/mjeanroy/rest_assert/internal/data/bindings/JavaxCookie.java
@@ -84,7 +84,7 @@ public class JavaxCookie implements Cookie {
 	}
 
 	@Override
-	public long getMaxAge() {
-		return cookie.getMaxAge();
+	public Long getMaxAge() {
+		return (long) cookie.getMaxAge();
 	}
 }
